refactor: フィールドが再レンダリングされないようリファクタ
